+++ project/routes/user_routes.py
@@ -15,6 +15,8 @@ blueprint = Blueprint(
 # Get User List Route
 @blueprint.route('/')
 def users():
+    if not g.admin:
+        return redirect(url_for('auth.login'))
     page = request.args.get('page', type=int, default=1)
     needle = request.args.get('needle', type=str, default='')
     user_list = UserModel.query.order_by(UserModel.registered_on.desc())
@@ -28,6 +30,8 @@ def users():
 # Register User Route
 @blueprint.route('/register', methods=['POST'])
 def register():
+    if not g.admin:
+        return redirect(url_for('auth.login'))
     user_name = request.form['user_name']
     user = UserModel(user_name=user_name)
     g.db.add(user)
@@ -37,6 +41,8 @@ def register():
 # Delete User Route
 @blueprint.route('/delete', methods=['POST'])
 def delete():
+    if not g.admin:
+        return redirect(url_for('auth.login'))
     user_id = request.form['user_id']
     user = UserModel.query.get_or_404(user_id)
     g.db.delete(user)
